+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Unreleased
 
+### Added
+
+- New trivia from
+  - _The Clone Wars_
+
 ## 0.13.0 - 2020-12-23
 
 ### Added